requirements.txt


# Core AI and ML libraries
openai>=1.3.0
anthropic>=0.8.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
llama-index>=0.9.0

# Vector databases and embeddings
chromadb>=0.4.18
pinecone-client>=2.2.4
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Traditional databases
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Web frameworks and APIs
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
gradio>=3.50.0
websockets>=12.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.6.0
jupyter>=1.0.0
jupyterlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0
requests>=2.31.0
httpx>=0.25.0
pyyaml>=6.0.1

# Monitoring and logging
loguru>=0.7.0
prometheus-client>=0.19.0
APScheduler>=3.10.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
EOF

# Install all dependencies
pip install -r requirements.txt

# Create requirements-dev.txt for additional development tools
cat > requirements-dev.txt << 'EOF'
# Development-only dependencies
ipython>=8.16.0
pre-commit>=3.5.0
bandit>=1.7.5
coverage>=7.3.0
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
EOF

pip install -r requirements-dev.txt